scse_data_archiving/requirements.txt
Django==1.6.5
Pillow==2.5.3
South==1.0
argparse==1.2.1
django-appconf==0.6
django-autoslug-field==0.2.3
django-braces==1.4.0
django-compressor==1.4
django-forms-bootstrap==3.0.0
django-grappelli==2.5.3
django-model-utils==2.0.3
django-reversion==1.8.1
django-sortedm2m==0.7.0
django-stdimage==1.0.2
django-suit==0.2.9
ipdb==0.8
ipython==2.1.0
psycopg2==2.5.3
pytz==2014.4
six==1.7.3
wsgiref==0.1.2
